# Suporte Service/src/main/resources/swagger/api.yml
swagger: "2.0"
info:
  description: '
  Trabalho de Conclusão da Disciplina de SERVICES ARCHITECTURE / API / MOBILE ARCHITECTURE<br><br>
  Case da aplicação Netflix<br><br>
  <b>Grupo:</b><br>
  Bruno Vinicius Cavalcante da Silva - 336968<br>
  Felipe Verissimo Gonzalez - 337283<br>
  Gabriel Muniz de Lima - 337821<br>
  '
  version: "1.0.0"
  title: "TCD - Netflix"
tags:
  - name: "suporte"
    description: "Relatar problemas que podem estar acontecendo."
  - name: "filmes"
    description: "Parte de filmes"
  - name: "usuario"
    description: "Parte de usuarios"
schemes:
  - "http"
paths:
  /suporte/reportar:
    post:
      tags:
        - "suporte"
      summary: "Cria um novo chamado para a equipe de suporte."
      operationId: "suporteReportarPost"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Informações do problema ocorrido"
          required: true
          schema:
            $ref: "#/definitions/Problema"
      responses:
        "200":
          description: "Ok Message"
          schema:
            type: "string"
        "405":
          description: "Invalid input"
  /suporte/consultarByDesc:
    get:
      tags:
        - "suporte"
      summary: "Consulta chamados que contenham o texto."
      operationId: "suporteConsultarByDescGet"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "texto"
          description: "Trecho de texto para consulta."
          required: true
          type: "string"
      responses:
        "200":
          description: "Lista de problemas."
          schema:
            type: "array"
            items:
              $ref: '#/definitions/TicketConsulta'
        "405":
          description: "Invalid input"
definitions:
  Problema:
    type: "object"
    properties:
      idFilme:
        type: "integer"
        format: "int64"
      descricaoErro:
        type: "string"
        description: "Descrição do problema"
      idUsuario:
        type: "integer"
        format: "int64"
        description: "Id do usuário que esta relatando o problema."
  TicketConsulta:
    type: "object"
    properties:
      idTicket:
        type: "integer"
        format: "int64"
      idFilme:
        type: "integer"
        format: "int64"
        description: "Id do filme em que ocorreu o problema."
      idUsuario:
        type: "integer"
        format: "int64"
        description: "Id do usuário que esta relatando o problema."
      descricaoErro:
        type: "string"
        description: "Descrição do problema."
      dataCriacao:
        type: "string"
        description: "Data de criação."
      idAtendente:
        type: "integer"
        format: "int64"
        description: "Id do atendente para resolver o problema."